+++ src/overloads/index.ts
@@ -1,7 +1,9 @@
+import { config } from './config';
 import { switchCmd } from './switch';
 
 const commands: Record<string, () => void | Promise<void>> = {
   switch: switchCmd,
+  config,
 };
 
 /** Run an custom workflow for a git/gh command if no args are passed in */